signal_processing/bias_triangle_processing/requirements.txt
numpy~=1.23.3
matplotlib~=3.5.2
opencv-python-headless~=4.7
scikit-image~=0.19.2
scipy~=1.9.3
jupytext~=1.14.6
tqdm~=4.65.0
xarray~=2023.5.0
black~=23.3.0
netcdf4~=1.6.4
mlflow~=2.4.1
mindfoundry_optaas_client~=1.4.1
# add this when installing --extra-index-url="https://opt-pre-release.internal.mindfoundry.ai/static/client/mindfoundry_optaas_client-1.beta-py3-none-any.whl"
gpytorch~=1.11
pytest~=7.4.0
qcodes
mpl_image_labeller~=1.1.2
fastai~=2.7.12